# docs/openapi.yaml
openapi: 3.0.3
info:
  title: India In-Time API
  version: "3.6.0"
  description: |
    Production API contract for India In-Time. Request and response schemas are
    intentionally explicit so frontend, mobile, QA, and third-party clients can
    validate payloads without reverse-engineering route implementations.
servers:
  - url: /api
    description: Deployment-relative API root
...
paths:
...
  /geocode:
    get:
      tags: [Geocode]
      summary: Geocode a place name
      parameters:
        - { $ref: '#/components/parameters/QueryQ' }
      responses:
        '200':
          description: Geocoding result set
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/GeocodeResult' }
                required: [results]
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /places:
    get:
      tags: [Places]
      summary: Discover places for a city
      parameters:
        - { name: city, in: query, required: false, schema: { type: string, maxLength: 100 } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200':
          description: Place list
          content:
            application/json:
              schema:
                type: object
                properties:
                  city: { type: string }
                  places:
                    type: array
                    items: { $ref: '#/components/schemas/Place' }
                required: [places]
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /weather:
    get:
      tags: [Weather]
      summary: Weather proxy
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number, minimum: -90, maximum: 90 } }
        - { name: lon, in: query, required: true, schema: { type: number, minimum: -180, maximum: 180 } }
      responses:
        '200':
          description: Weather payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WeatherResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /ai:
    post:
      tags: [AI]
      summary: Generate an AI response
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AiRequest' }
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502': { $ref: '#/components/responses/BadGateway' }
  /trips:
    get:
      tags: [Trips]
      summary: List current user's saved trips
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200':
          description: Saved trips
          content:
            application/json:
              schema:
                type: object
                properties:
                  trips:
                    type: array
                    items: { $ref: '#/components/schemas/Trip' }
                required: [trips]
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Trips]
      summary: Save a trip
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TripCreate' }
      responses:
        '201':
          description: Trip created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Trip' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /favorites:
    get:
      tags: [Favorites]
      summary: List current user's active favorites
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorites:
                    type: array
                    items: { $ref: '#/components/schemas/Favorite' }
                required: [favorites]
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Favorites]
      summary: Add or restore a favorite
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FavoriteCreate' }
      responses:
        '201':
          description: Favorite created or restored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Favorite' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  parameters:
    QueryQ:
      name: q
      in: query
      required: true
      schema: { type: string, minLength: 2, maxLength: 120 }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    BadGateway:
      description: Upstream provider unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        requestId: { type: string }
...
    Coordinates:
      type: array
      minItems: 2
      maxItems: 2
      items: { type: number }
      example: [17.6868, 83.2185]
    GeocodeResult:
      type: object
      properties:
        display_name: { type: string }
        lat: { type: number }
        lon: { type: number }
        type: { type: string }
    Place:
      type: object
      required: [id, name, category, coords]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        coords: { $ref: '#/components/schemas/Coordinates' }
        visitTimeMinutes: { type: integer, minimum: 0 }
        openTime: { type: string, pattern: '^\\d{2}:\\d{2}$' }
        closeTime: { type: string, pattern: '^\\d{2}:\\d{2}$' }
        rating: { type: number, minimum: 0, maximum: 5 }
        importance: { type: string }
    WeatherResponse:
      type: object
      properties:
        temperature: { type: number }
        feelsLike: { type: number }
        humidity: { type: number, minimum: 0, maximum: 100 }
        weather: { type: string }
        description: { type: string }
        windSpeed: { type: number }
        observedAt: { type: string, format: date-time }
...
    AiRequest:
      type: object
      required: [prompt]
      properties:
        prompt: { type: string, minLength: 1, maxLength: 10000 }
        system: { type: string, maxLength: 5000 }
        model: { type: string, maxLength: 100 }
        temperature: { type: number, minimum: 0, maximum: 2, default: 0.4 }
        maxOutputTokens: { type: integer, minimum: 1, maximum: 8192, default: 2048 }
    AiResponse:
      type: object
      required: [text]
      properties:
        text: { type: string }
        model: { type: string }
        cached: { type: boolean }
        requestId: { type: string }
    TripCreate:
      type: object
      required: [city, plan]
      properties:
        city: { type: string, minLength: 1, maxLength: 100 }
        startTime: { type: string, pattern: '^\\d{2}:\\d{2}$' }
        durationMinutes: { type: integer, minimum: 1, maximum: 1440 }
        plan:
          type: array
          maxItems: 30
          items:
            type: array
            items: { $ref: '#/components/schemas/TripStop' }
        budget: { type: number, minimum: 0 }
        metadata: { type: object, additionalProperties: true }
    Trip:
      allOf:
        - $ref: '#/components/schemas/TripCreate'
        - type: object
          required: [id]
          properties:
            id: { type: string }
            userId: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    TripStop:
      type: object
      required: [name]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        coords: { $ref: '#/components/schemas/Coordinates' }
        startTime: { type: string }
        endTime: { type: string }
        visitMinutes: { type: integer, minimum: 0 }
    FavoriteCreate:
      type: object
      required: [placeName, city]
      properties:
        placeName: { type: string, minLength: 1, maxLength: 200 }
        city: { type: string, minLength: 1, maxLength: 100 }
        category: { type: string, maxLength: 50 }
        coords: { $ref: '#/components/schemas/Coordinates' }
        notes: { type: string, maxLength: 2000 }
    Favorite:
      allOf:
        - $ref: '#/components/schemas/FavoriteCreate'
        - type: object
          required: [id]
          properties:
            id: { type: integer, format: int64 }
            createdAt: { type: string, format: date-time }
...